refactor: filtered table withoutexternal generic component
