+++ src/components/standalone/firewall/conntrack/DeleteConntrackRecordModal.vue
@@ -77,8 +77,8 @@ function close() {
     <span v-if="itemToDelete?.id !== 'all'">
       {{
         t('standalone.conntrack.delete_conntrack_record_message', {
-          source: itemToDelete?.destination ?? '',
-          destination: itemToDelete?.source ?? ''
+          source: itemToDelete?.destination + ':' + itemToDelete?.destination_port ?? '',
+          destination: itemToDelete?.source + ':' + itemToDelete?.source_port ?? ''
         })
       }}
     </span>
